Add set of Selection on BBP config reader object (from Neuroscheme)
Tidy up listening of events from zeq and the init/cleanup of the zeq receiver.

vtkZeqManager spawns a listening thread and then sends received events to
GUI via a TCP socket which is receiving in the pq GUI control.
The GUI then sends the received GIds from a selction on to any BBP
circuit reader classes that might be instantiated.