Improve forwarding of messages from server to client
There is still a bug that messages are sometimes not fully
received when we try to read them from the socket.

Unified the handling of messages from the server with the handling
from the client so that either client or server can listen for zeq events.

Flyby fixes to segfaults on server when starting up.